Move options to VariantEvalArgumentCollection, to allow setting values more easily by AlleleFrequencyQC
